+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## [0.1.0] - 12/11/2025
 
 ### Added
+- Consult Product Batches + Register ViewModel and Screen
 - DeleteOneSupply Button Logic
 - Filter and search bar for users
 - US to add a new supply with an image